airflow/requirements.txt
aenum==3.1.15
agate==1.9.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
apache-airflow-providers-celery==3.10.0
apache-airflow-providers-common-compat==1.2.0
apache-airflow-providers-common-io==1.4.0
apache-airflow-providers-common-sql==1.24.0
apache-airflow-providers-fab==1.3.0
apache-airflow-providers-ftp==3.11.0
apache-airflow-providers-http==4.13.0
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-postgres==5.14.0
apache-airflow-providers-slack==9.0.2
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.6.1
argcomplete==3.5.0
asgiref==3.8.1
astronomer-cosmos==1.9.1
attrs==24.2.0
babel==2.16.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.8.2
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
daff==1.3.46
dbt-adapters==1.14.0
dbt-common==1.15.0
dbt-core==1.9.2
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distlib==0.3.9
dnspython==2.6.1
docutils==0.16
email_validator==2.2.0
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.0
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
flower==2.0.1
frozenlist==1.4.1
fsspec==2024.9.0
google-re2==1.1.20240702
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.12.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.5.0
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
ordered-set==4.1.0
packaging==24.1
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.8.1
rich-argparse==1.5.2
rpds-py==0.20.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
slack_sdk==3.34.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.15.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.5
urllib3==2.2.2
uv==0.4.1
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.11.1
zipp==3.20.2